refactor(player): extract sections from GameDetailScreen (#695)
`GameDetailScreen.kt` was 1094 LOC mixing the screen-level layout
glue with several private composables that were big enough to own
their own files in the existing `feature/gamedetail/` folder.

Extracted to `feature/gamedetail/`:

- **GameHeroContent.kt** (309 LOC) — the hero banner content:
  title, badges, ratings, action buttons, status indicators,
  playtime chips. Plus the two private rating-badge helpers
  (IgdbRatingStars + CommunityRatingBadge) that were only used
  by it.
- **VariantsSection.kt** (170 LOC) — three sibling sections that
  surface related ROMs (alternate Versions, Based On, ROM Hacks).
  Same SpCard layout, only header label + data source differ.
- **GameInfoContent.kt** (157 LOC) — the middle column:
  description, MetadataGrid call, variants splitting, BasedOn,
  RomHacks. Pure orchestration of the above.

LOC:
- GameDetailScreen.kt: 1094 → 478 (-616, -56%)
- New extracted composables: 636 LOC
- Net codebase delta: +20 LOC, but the screen file is now mostly
  the top-level GameDetailLayout call + the loading / scaffolding
  / dialogs glue. Each extracted piece lives in the feature
  folder where game-detail features already cluster (alongside
  GameActionsMenu, MetadataGrid, GameSharedSessionsSection, etc).

Imports trimmed via detekt's UnusedImport report (46 stale
imports removed from the screen file after the extraction).

Behaviour preserved. 526 shared + 759 desktop tests pass. Both
targets compile (desktop + Android). detektMainDesktop clean.

Closes #695
