+++ player/shared/src/commonMain/kotlin/com/spela/player/presentation/ui/screen/GameDetailScreen.kt
@@ -1,61 +1,26 @@
 package com.spela.player.presentation.ui.screen
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Text
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import com.spela.player.domain.model.BiosMissingFile
-import com.spela.player.domain.model.DownloadState
-import com.spela.player.domain.model.Game
-import com.spela.player.domain.model.GameDetail
-import com.spela.player.domain.model.GameVariant
-import com.spela.player.domain.model.ParentGame
-import com.spela.player.domain.model.RomHackGame
-import com.spela.player.domain.model.NETPLAY_SUPPORTED_CONSOLES
 import com.spela.player.presentation.intent.GameDetailIntent
-import com.spela.player.presentation.state.GameDetailState
 import com.spela.player.presentation.state.GameSyncState
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AccessTime
-import androidx.compose.material.icons.filled.History
-import androidx.compose.material.icons.filled.EmojiEvents
 import androidx.compose.material.icons.filled.Star
-import androidx.compose.material3.Icon
 import com.spela.player.presentation.ui.feature.collections.CollectionPickerDialog
 import com.spela.player.presentation.intent.KeyMappingIntent
-import com.spela.player.presentation.ui.feature.gamedetail.BiosWarningChip
-import com.spela.player.presentation.ui.feature.gamedetail.GameActionsMenu
-import com.spela.player.presentation.ui.feature.gamedetail.MetadataGrid
-import com.spela.player.presentation.ui.feature.gamedetail.VerificationChip
+import com.spela.player.presentation.ui.feature.gamedetail.GameHeroContent
+import com.spela.player.presentation.ui.feature.gamedetail.GameInfoContent
 import com.spela.player.presentation.ui.feature.gamedetail.ChallengesSection
 import com.spela.player.presentation.ui.feature.gamedetail.CreateChallengeDialog
 import com.spela.player.presentation.ui.feature.gamedetail.CommunitySharesSection
@@ -70,35 +35,22 @@ import com.spela.player.presentation.ui.feature.gamedetail.TimeToBeatSection
 import com.spela.player.presentation.ui.feature.gamedetail.ScreenshotsSection
 import com.spela.player.presentation.ui.feature.gamedetail.SimilarGamesSection
 import com.spela.player.presentation.ui.feature.library.darken
-import com.spela.player.presentation.ui.feature.library.getConsoleColor
 import com.spela.player.presentation.ui.feature.library.getConsoleGradient
 import com.spela.player.presentation.ui.components.GameDetailLayout
 import com.spela.player.presentation.ui.components.GameDetailSkeleton
-import com.spela.player.presentation.ui.components.SpButton
-import com.spela.player.presentation.ui.components.SpButtonStyle
-import com.spela.player.presentation.ui.components.SpCard
-import com.spela.player.presentation.ui.components.SpChip
-import com.spela.player.presentation.ui.components.SpRegionChip
 import com.spela.player.presentation.ui.components.SpConfirmDialog
-import com.spela.player.presentation.ui.components.SpConsoleChip
 import com.spela.player.presentation.ui.components.SpCoverArt
 import com.spela.player.presentation.ui.components.SpSnackbar
 import com.spela.player.presentation.ui.components.SpSnackbarData
 import com.spela.player.presentation.ui.components.SpSnackbarType
-import com.spela.player.presentation.ui.components.SpSplitButton
-import com.spela.player.presentation.ui.components.SpSplitButtonMenuItem
 import com.spela.player.presentation.ui.components.SpTitledSection
 import com.spela.player.presentation.ui.components.SpTopBar
-import com.spela.player.presentation.ui.gamepad.autoFocus
 import com.spela.player.presentation.ui.components.PlatformBackHandler
 import com.spela.player.presentation.ui.components.social.StarRatingRow
-import com.spela.player.presentation.ui.components.social.formatRelativeTime
 import com.spela.player.presentation.ui.theme.SpColor
-import com.spela.player.presentation.ui.theme.SpSpacing
 import com.spela.player.presentation.ui.theme.SpTypography
 import com.spela.player.presentation.viewmodel.GameDetailViewModel
 import com.spela.player.presentation.viewmodel.KeyMappingViewModel
-import com.spela.player.util.formatPlayTime
 
 @Composable
 fun GameDetailScreen(
@@ -522,573 +474,5 @@ fun GameDetailScreen(
             onDismiss = { viewModel.onIntent(GameDetailIntent.DismissError) },
             modifier = Modifier.align(Alignment.BottomCenter),
         )
-    }
-}
-
-@Composable
-private fun GameInfoContent(
-    gameId: String,
-    game: Game,
-    detail: GameDetail,
-    state: GameDetailState,
-    isPortrait: Boolean = false,
-    hasSaves: Boolean,
-    missingBiosFiles: List<BiosMissingFile> = emptyList(),
-    isDemoConsole: Boolean = false,
-    onPlay: (String) -> Unit,
-    onPlayFresh: ((String) -> Unit)? = null,
-    onPlayFromTitleScreen: ((String) -> Unit)? = null,
-    onDownloadGame: () -> Unit,
-    onToggleFavorite: () -> Unit,
-    onTogglePlayLater: () -> Unit,
-    onAddToCollection: () -> Unit,
-    onCreateNetplay: ((String) -> Unit)? = null,
-    onDeleteLocalGame: () -> Unit = {},
-    onRate: (Int) -> Unit = {},
-    syncState: GameSyncState? = null,
-    onPlayWithLocalSave: () -> Unit = {},
-    onCancelLaunch: () -> Unit = {},
-    onNavigateToGame: ((String) -> Unit)? = null,
-    onNavigateToDeveloper: ((name: String) -> Unit)? = null,
-    onNavigateToPublisher: ((name: String) -> Unit)? = null,
-    onNavigateToAchievements: (() -> Unit)? = null,
-) {
-    // Title, badges, action buttons, and loading indicators are in the hero banner (GameHeroContent)
-
-    // BIOS warning chip
-    if (missingBiosFiles.isNotEmpty()) {
-        var showBiosInfo by remember { mutableStateOf(false) }
-
-        BiosWarningChip(
-            missingFiles = missingBiosFiles,
-            onClick = { showBiosInfo = !showBiosInfo },
-        )
-        AnimatedVisibility(visible = showBiosInfo) {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = SpSpacing.Small)
-                    .background(
-                        SpColor.Warning.copy(alpha = 0.1f),
-                        RoundedCornerShape(SpSpacing.RadiusMedium),
-                    )
-                    .padding(SpSpacing.Medium)
-                    .semantics { contentDescription = "Missing BIOS files info" },
-            ) {
-                Text(
-                    text = "Missing BIOS files:",
-                    style = SpTypography.LabelMedium,
-                    color = SpColor.Warning,
-                )
-                missingBiosFiles.forEach { file ->
-                    Text(
-                        text = file.fileName,
-                        style = SpTypography.BodySmall,
-                        color = SpColor.OnBackgroundSecondary,
-                    )
-                }
-            }
-        }
-    }
-
-    // Description (plain text, matching web UI)
-    game.description?.let { description ->
-        Text(
-            text = description,
-            style = SpTypography.BodyMedium,
-            color = SpColor.OnBackgroundSecondary,
-        )
-    }
-
-    // Metadata grid (Developer, Publisher, Released, Genre, Players, Achievements, Size, Discs)
-    MetadataGrid(
-        game = game,
-        onGradient = true,
-        isDemoConsole = isDemoConsole,
-        achievementTotal = state.achievements.size,
-        achievementUnlocked = state.achievementProgress.size,
-        onDeveloperClick = onNavigateToDeveloper,
-        onPublisherClick = onNavigateToPublisher,
-        onAchievementsClick = onNavigateToAchievements,
-    )
-
-    // Variants section -- split into Versions (non-hack) and ROM Hacks (hack-tagged)
-    val versionVariants = detail.variants.filter { variant ->
-        variant.tags?.split(",")?.map { it.trim().lowercase() }?.contains("hack") != true
-    }
-    val hackVariants = detail.variants.filter { variant ->
-        variant.tags?.split(",")?.map { it.trim().lowercase() }?.contains("hack") == true
-    }
-
-    if (versionVariants.isNotEmpty()) {
-        Spacer(Modifier.height(SpSpacing.Default))
-        VariantsSection(
-            title = "Versions",
-            variants = versionVariants,
-            onVariantSelected = onNavigateToGame,
-        )
-    }
-
-    if (hackVariants.isNotEmpty()) {
-        Spacer(Modifier.height(SpSpacing.Default))
-        VariantsSection(
-            title = "ROM Hacks",
-            variants = hackVariants,
-            onVariantSelected = onNavigateToGame,
-        )
-    }
-
-    // "Based on" section for standalone ROM hacks
-    detail.parentGame?.let { parent ->
-        Spacer(Modifier.height(SpSpacing.Default))
-        BasedOnSection(
-            parentGame = parent,
-            onNavigateToGame = onNavigateToGame,
-        )
-    }
-
-    // Standalone ROM Hacks section
-    if (detail.romHacks.isNotEmpty()) {
-        Spacer(Modifier.height(SpSpacing.Default))
-        RomHacksSection(
-            romHacks = detail.romHacks,
-            onNavigateToGame = onNavigateToGame,
-        )
-    }
-}
-
-@Composable
-private fun IgdbRatingStars(rating: Double) {
-    val normalized = rating / 10.0
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(SpSpacing.XXSmall),
-        modifier = Modifier.semantics {
-            contentDescription = "IGDB rating: ${"%.1f".format(normalized)} out of 10"
-        },
-    ) {
-        Icon(
-            imageVector = Icons.Filled.Star,
-            contentDescription = null,
-            tint = SpColor.Warning,
-            modifier = Modifier.size(16.dp),
-        )
-        Text(
-            text = "%.1f".format(normalized),
-            style = SpTypography.LabelMedium,
-            color = SpColor.OnBackgroundSecondary,
-        )
-    }
-}
-
-@Composable
-private fun CommunityRatingBadge(averageRating: Double, ratingCount: Long) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(SpSpacing.XXSmall),
-        modifier = Modifier.semantics {
-            contentDescription = "Community rating: ${"%.1f".format(averageRating)} from $ratingCount ratings"
-        },
-    ) {
-        Icon(
-            imageVector = Icons.Filled.Star,
-            contentDescription = null,
-            tint = SpColor.Warning,
-            modifier = Modifier.size(20.dp),
-        )
-        Text(
-            text = "%.1f".format(averageRating),
-            style = SpTypography.LabelMedium,
-            color = SpColor.OnBackgroundTertiary,
-        )
-        Text(
-            text = "($ratingCount)",
-            style = SpTypography.LabelMedium,
-            color = SpColor.OnBackgroundTertiary,
-        )
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun VariantsSection(
-    title: String,
-    variants: List<GameVariant>,
-    onVariantSelected: ((String) -> Unit)?,
-) {
-    SpTitledSection(
-        title = title,
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-        ) {
-            variants.forEach { variant ->
-                SpCard(
-                    onClick = { onVariantSelected?.invoke(variant.id) },
-                    onGradient = true,
-                    cornerRadius = SpSpacing.RadiusMedium,
-                ) {
-                    Column(
-                        modifier = Modifier.fillMaxWidth().padding(SpSpacing.Medium),
-                        verticalArrangement = Arrangement.spacedBy(SpSpacing.XSmall),
-                    ) {
-                        Text(
-                            text = variant.title,
-                            style = SpTypography.TitleSmall,
-                            color = SpColor.OnCard,
-                        )
-                        FlowRow(
-                            horizontalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-                            verticalArrangement = Arrangement.spacedBy(SpSpacing.XSmall),
-                        ) {
-                            variant.region?.takeIf { it.isNotBlank() }?.let { region ->
-                                SpRegionChip(region = region, onGradient = true)
-                            }
-                            variant.revision?.takeIf { it.isNotBlank() }?.let { revision ->
-                                SpChip(text = revision, onGradient = true)
-                            }
-                            if (variant.fileSize > 0) {
-                                SpChip(
-                                    text = formatVariantFileSize(variant.fileSize),
-                                    onGradient = true,
-                                )
-                            }
-                            VerificationChip(
-                                verificationStatus = variant.verificationStatus,
-                            )
-                            if (variant.isPreRelease) {
-                                SpChip(text = "Pre-release", onGradient = true)
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun formatVariantFileSize(bytes: Long): String {
-    if (bytes < 1024) return "$bytes B"
-    val kb = bytes / 1024.0
-    if (kb < 1024) return String.format("%.0f KB", kb)
-    val mb = kb / 1024.0
-    if (mb < 1024) return String.format("%.1f MB", mb)
-    val gb = mb / 1024.0
-    return String.format("%.2f GB", gb)
-}
-
-@Composable
-private fun BasedOnSection(
-    parentGame: ParentGame,
-    onNavigateToGame: ((String) -> Unit)?,
-) {
-    SpTitledSection(
-        title = "Based on",
-    ) {
-        SpCard(
-            onClick = { onNavigateToGame?.invoke(parentGame.id) },
-            onGradient = true,
-            cornerRadius = SpSpacing.RadiusMedium,
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth().padding(SpSpacing.Medium),
-                horizontalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                SpCoverArt(
-                    imageUrl = parentGame.coverUrl,
-                    contentDescription = "${parentGame.title} cover art",
-                    modifier = Modifier.size(width = SpSpacing.CoverSmallWidth, height = SpSpacing.CoverSmallHeight),
-                )
-                Text(
-                    text = parentGame.title,
-                    style = SpTypography.TitleSmall,
-                    color = SpColor.OnCard,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun RomHacksSection(
-    romHacks: List<RomHackGame>,
-    onNavigateToGame: ((String) -> Unit)?,
-) {
-    SpTitledSection(
-        title = "ROM Hacks",
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-        ) {
-            romHacks.forEach { hack ->
-                SpCard(
-                    onClick = { onNavigateToGame?.invoke(hack.id) },
-                    onGradient = true,
-                    cornerRadius = SpSpacing.RadiusMedium,
-                ) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth().padding(SpSpacing.Medium),
-                        horizontalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        SpCoverArt(
-                            imageUrl = hack.coverUrl,
-                            contentDescription = "${hack.title} cover art",
-                            modifier = Modifier.size(width = SpSpacing.CoverSmallWidth, height = SpSpacing.CoverSmallHeight),
-                        )
-                        Text(
-                            text = hack.title,
-                            style = SpTypography.TitleSmall,
-                            color = SpColor.OnCard,
-                            maxLines = 2,
-                            overflow = TextOverflow.Ellipsis,
-                            modifier = Modifier.weight(1f),
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Content rendered inside the hero banner's contrast backdrop in portrait mode.
- * Shows game title, badges, ratings, and action buttons.
- */
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun GameHeroContent(
-    gameId: String,
-    game: Game,
-    state: GameDetailState,
-    hasSaves: Boolean,
-    missingBiosFiles: List<BiosMissingFile>,
-    onPlay: (String) -> Unit,
-    onPlayFresh: ((String) -> Unit)?,
-    onPlayFromTitleScreen: ((String) -> Unit)? = null,
-    onDownloadGame: () -> Unit,
-    onToggleFavorite: () -> Unit,
-    onTogglePlayLater: () -> Unit,
-    onAddToCollection: () -> Unit,
-    onCreateNetplay: ((String) -> Unit)?,
-    onDeleteLocalGame: () -> Unit,
-    syncState: GameSyncState?,
-    onNavigateToAchievements: () -> Unit = {},
-    onAdminScrape: (() -> Unit)? = null,
-    onAdminRefreshAchievements: (() -> Unit)? = null,
-) {
-    val supportsNetplay = game.playable && game.consoleId.lowercase() in NETPLAY_SUPPORTED_CONSOLES
-
-    Column(
-        verticalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-    ) {
-        // Title
-        Text(
-            text = game.title,
-            style = SpTypography.DisplaySmall,
-            fontWeight = FontWeight.Bold,
-            color = Color.White,
-            maxLines = 2,
-            overflow = TextOverflow.Ellipsis,
-        )
-
-        // Badges: console, region, verification, ratings
-        FlowRow(
-            horizontalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-            verticalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-            itemVerticalAlignment = Alignment.CenterVertically,
-        ) {
-            SpConsoleChip(
-                consoleName = game.consoleName,
-                consoleColor = getConsoleColor(game.consoleName),
-                onGradient = true,
-            )
-            game.region?.takeIf { it.isNotBlank() }?.let { region ->
-                SpRegionChip(region = region, onGradient = true)
-            }
-            VerificationChip(
-                verificationStatus = game.verificationStatus,
-                verificationTag = game.verificationTag,
-            )
-            if (game.igdbCriticsRating > 0) {
-                IgdbRatingStars(rating = game.igdbCriticsRating)
-            }
-            if (game.communityRating > 0) {
-                CommunityRatingBadge(
-                    averageRating = game.communityRating,
-                    ratingCount = game.communityRatingCount,
-                )
-            }
-            if (state.achievements.isNotEmpty()) {
-                SpChip(
-                    text = "${state.achievementProgress.size} / ${state.achievements.size}",
-                    leadingIcon = {
-                        Icon(
-                            imageVector = Icons.Filled.EmojiEvents,
-                            contentDescription = null,
-                            tint = Color.White.copy(alpha = 0.65f),
-                            modifier = Modifier.size(14.dp),
-                        )
-                    },
-                    onGradient = true,
-                    onClick = onNavigateToAchievements,
-                )
-            }
-        }
-
-        // Action buttons: Play/Download + Actions menu + playtime
-        FlowRow(
-            horizontalArrangement = Arrangement.spacedBy(SpSpacing.Small),
-            verticalArrangement = Arrangement.spacedBy(SpSpacing.Medium),
-            itemVerticalAlignment = Alignment.CenterVertically,
-        ) {
-            if (state.isGameCached) {
-                if (game.playable) {
-                    val menuItems = buildList {
-                        if (hasSaves && onPlayFromTitleScreen != null) {
-                            add(SpSplitButtonMenuItem(
-                                label = "Continue from Title Screen",
-                                description = "Keep your in-game save, start from the beginning",
-                            ) { onPlayFromTitleScreen(gameId) })
-
-                        }
-                        if (hasSaves && onPlayFresh != null) {
-                            add(SpSplitButtonMenuItem(
-                                label = "New Game",
-                                description = "Start a separate playthrough from scratch",
-                            ) { onPlayFresh(gameId) })
-                        }
-                        if (onCreateNetplay != null && supportsNetplay) {
-                            add(SpSplitButtonMenuItem("Netplay") { onCreateNetplay(gameId) })
-                        }
-                        add(SpSplitButtonMenuItem("Delete Download") { onDeleteLocalGame() })
-                    }
-                    val hasRequiredBiosMissing = missingBiosFiles.any { it.required }
-                    val isSyncing = syncState != null
-                    SpSplitButton(
-                        text = if (hasSaves) "Resume" else "Play",
-                        onClick = { onPlay(gameId) },
-                        modifier = Modifier
-                            .autoFocus()
-                            .testTag("game_detail_play_button"),
-                        enabled = !hasRequiredBiosMissing && !isSyncing,
-                        isLoading = false,
-                        menuItems = menuItems,
-                        onGradient = true,
-                    )
-                } else {
-                    SpButton(
-                        text = "Delete Download",
-                        onClick = onDeleteLocalGame,
-                        style = SpButtonStyle.Ghost,
-                        onGradient = true,
-                    )
-                }
-            } else {
-                val isActivelyDownloading = state.downloadProgress?.state == DownloadState.DOWNLOADING
-                val isBusy = state.isDownloading || isActivelyDownloading
-                val menuItems = buildList {
-                    if (onCreateNetplay != null && supportsNetplay) {
-                        add(SpSplitButtonMenuItem("Netplay") { onCreateNetplay(gameId) })
-                    }
-                }
-                SpSplitButton(
-                    text = if (isBusy) "Downloading..." else "Download",
-                    onClick = onDownloadGame,
-                    modifier = Modifier.testTag("game_detail_download_button"),
-                    isLoading = isBusy,
-                    enabled = !isBusy,
-                    menuItems = menuItems,
-                    onGradient = true,
-                )
-            }
-
-            GameActionsMenu(
-                isFavorite = game.isFavorite,
-                isInPlayLater = game.isInPlayLater,
-                onToggleFavorite = onToggleFavorite,
-                onTogglePlayLater = onTogglePlayLater,
-                onAddToCollection = onAddToCollection,
-                onGradient = true,
-                onAdminScrape = onAdminScrape,
-                onAdminRefreshAchievements = onAdminRefreshAchievements,
-                isAdminActionLoading = state.isAdminActionLoading,
-            )
-
-            // Status indicators (scraping, syncing, downloading)
-            val statusText = when {
-                state.isScraping -> "Scraping\u2026"
-                syncState != null && !syncState.isTimedOut -> syncState.message
-                state.isDownloading -> {
-                    val p = state.downloadProgress
-                    if (p != null && p.totalDiscs > 1) "Downloading disc ${p.currentDisc}/${p.totalDiscs}\u2026"
-                    else "Downloading\u2026"
-                }
-                else -> null
-            }
-            if (statusText != null) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(SpSpacing.XSmall),
-                ) {
-                    // Show spinner for non-download statuses (scraping, sync).
-                    // Downloads already have a spinner in the button.
-                    if (!state.isDownloading) {
-                        CircularProgressIndicator(
-                            modifier = Modifier.size(14.dp),
-                            strokeWidth = 2.dp,
-                            color = Color.White.copy(alpha = 0.65f),
-                        )
-                    }
-                    Text(
-                        text = statusText,
-                        style = SpTypography.LabelSmall,
-                        color = Color.White.copy(alpha = 0.65f),
-                    )
-                }
-            }
-
-            // Playtime + last played grouped so they wrap together
-            if (game.totalPlayTime > 0 || game.lastPlayedAt != null) {
-                Row(horizontalArrangement = Arrangement.spacedBy(SpSpacing.Small)) {
-                    if (game.totalPlayTime > 0) {
-                        SpChip(
-                            text = formatPlayTime(game.totalPlayTime),
-                            onGradient = true,
-                            leadingIcon = {
-                                Icon(
-                                    imageVector = Icons.Filled.AccessTime,
-                                    contentDescription = null,
-                                    tint = Color.White.copy(alpha = 0.65f),
-                                    modifier = Modifier.size(14.dp),
-                                )
-                            },
-                        )
-                    }
-                    game.lastPlayedAt?.let { timestamp ->
-                        val relative = formatRelativeTime(timestamp)
-                        if (relative.isNotEmpty()) {
-                            SpChip(
-                                text = relative,
-                                onGradient = true,
-                                leadingIcon = {
-                                    Icon(
-                                        imageVector = Icons.Filled.History,
-                                        contentDescription = null,
-                                        tint = Color.White.copy(alpha = 0.65f),
-                                        modifier = Modifier.size(14.dp),
-                                    )
-                                },
-                            )
-                        }
-                    }
-                }
-            }
-        }
     }
 }
